feat(extension): Working on local storage
Working on storing the group name and the user name locally using the
chrome's storage.sync API. This will allows the user to join the group
they were previously part of in their last session. This will avoid
the need to rejoin the group, everytime the extension is opened.

Commented out the ClipboardAPI. Will activate it after the storage is
working properly.
-----------------------------------------------------------------------
fix(extension): Able to load data successfully from storage API

Fixed the issue where the data was being stored correctly but not loaded back.
This was due to the time taken to load the data from the storage.sync API.
Add a timeout callback function after loading the data but before actually
using the values.