feat(extension): Working on local storage
Working on storing the group name and the user name locally using the
chrome's storage.sync API. This will allows the user to join the group
they were previously part of in their last session. This will avoid
the need to rejoin the group, everytime the extension is opened.

Commented out the ClipboardAPI. Will activate it after the storage is
working properly.